Added new mappings.

+++ backend/src/main/java/com/yukon/logistics/service/impl/CustomerServiceImpl.java
@@ -45,6 +45,11 @@ public class CustomerServiceImpl implements CustomerService {
 
     @Override
     public void deleteCustomerById(Long id) {
-        customerRepository.deleteById(id);
+        if(customerRepository.findById(id).isPresent()){
+            customerRepository.deleteById(id);
+        }
+        else{
+            throw new EntityNotFoundException(id + "Customer doesn't exist");
+        }
     }
 }
